canvas manager implementation

+++ Assets/Content/Stats/StatModifier.cs
@@ -8,13 +8,13 @@
 public class StatModifier
 {
     // readonly to make parameters be set only by constructor
-    public readonly float value;
+    public readonly int value;
     public readonly StatModType type;
     //object can hold any possible source type
     public readonly object source;
     public readonly int order;
 
-    public StatModifier(float value, StatModType type, object source, int order)
+    public StatModifier(int value, StatModType type, object source, int order)
     {
         this.value = value;
         this.type = type;
@@ -22,11 +22,11 @@ public class StatModifier
         this.order = order;
     }
 
-    public StatModifier(float value, StatModType type) : this(value, type, null, (int)type)
+    public StatModifier(int value, StatModType type) : this(value, type, null, (int)type)
     {
     }
 
-    public StatModifier(float value, StatModType type, object source) : this(value, type, source, (int)type)
+    public StatModifier(int value, StatModType type, object source) : this(value, type, source, (int)type)
     {
     }
